refactor: move display logic to a single client component and render page tsx on the server side to process user information
relates #393
